relocate collision package
