+++ cmd/rallymka/internal/stream/level.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/mokiat/gomath/sprec"
 	"github.com/mokiat/rally-mka/internal/data/asset"
-	"github.com/mokiat/rally-mka/internal/engine/collision"
 	"github.com/mokiat/rally-mka/internal/engine/graphics"
+	"github.com/mokiat/rally-mka/internal/engine/physics/collision"
 	"github.com/mokiat/rally-mka/internal/engine/resource"
 )
