+++ internal/engine/physics/collision.go
@@ -2,7 +2,7 @@ package physics
 
 import (
 	"github.com/mokiat/gomath/sprec"
-	"github.com/mokiat/rally-mka/internal/engine/collision"
+	"github.com/mokiat/rally-mka/internal/engine/physics/collision"
 )
 
 type BoxShape struct {
