+++ internal/engine/physics/collision/triangle.go
